Revert "Auth setup (#3)"
This reverts commit afa4fef2163528a7a38cfb8cad8bf41018b177e8.
