--- .eslintrc.js
@@ -1,19 +1,0 @@
-module.exports = {
-  env: {
-    browser: true,
-    commonjs: true,
-    es2021: true,
-  },
-  extends: 'airbnb-base',
-  overrides: [
-  ],
-  globals: {
-    i18n: true,
-  },
-  parserOptions: {
-    ecmaVersion: 'latest',
-  },
-  rules: {
-    'no-console': 'off',
-  },
-};
