--- src/models/index.js
@@ -1,7 +1,0 @@
-const Profile = require('./Profile');
-const User = require('./User');
-
-module.exports = {
-  Profile,
-  User,
-};
