--- src/utils/user.helper.js
@@ -1,13 +1,0 @@
-const bcrypt = require('bcryptjs');
-
-const hashpassword = async (password) => bcrypt.hash(password, 12);
-
-const comparePassword = async (reqPassword, dbPassword) => {
-  const result = bcrypt.compareSync(reqPassword, dbPassword);
-  return result;
-};
-
-module.exports = {
-  hashpassword,
-  comparePassword,
-};
